Add inherent `impl` block for `Mutex<RefCell<T>>`
Add inherent functions to `Mutex<RefCell<T>>` to reduce verbosity when
using the type. Add a note to the `Mutex` documentation explaining why
`Mutex::borrow` does not return `&mut T` and point out the newly created
methods on `Mutex<RefCell<T>>`. The design of `Mutex::borrow` is a
frequently asked question, so it helps to explicitly document the
decisions.